dancestrument/requirements.txt
absl-py==2.0.0
attrs==23.1.0
cffi==1.16.0
contourpy==1.1.1
cycler==0.12.1
flatbuffers==23.5.26
fonttools==4.43.1
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.8.0
mediapipe==0.10.7
numpy==1.26.1
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.1
Pillow==10.1.0
protobuf==3.20.3
pycparser==2.21
pyparsing==3.1.1
python-dateutil==2.8.2
python-osc==1.8.3
python-rtmidi==1.5.6
pytz==2023.3.post1
scikit-learn==1.3.1
scipy==1.11.3
six==1.16.0
sounddevice==0.4.6
threadpoolctl==3.2.0
tzdata==2023.3
